Improving the relationship and dependencies between Image Model entities.
Improving the attributes located in the ExamResult.java DTO class.
Creating the overrided equals method for the QuestionItem DTO class.
Improving the relationships between DTO in the SheetFileInformation.java class.
Comenting some source code lines in the MainTest.java.
Improving the via used for compare a pair of Question Items in the ExamSessionComparator.java.